tools/polymarket-monitor/requirements.txt
requests>=2.28.0
mysql-connector-python>=8.0.0
python-dotenv>=1.0.0
